+++ spec/www/spec/db-tx-sql-results.js
@@ -4,15 +4,9 @@ var MYTIMEOUT = 12000;
 
 var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 
-var isAndroid = /Android/.test(navigator.userAgent);
 var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
-//var isWindows = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-var isWindows = /Windows /.test(navigator.userAgent); // Windows (8.1)
-//var isWindowsPC = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-//var isWindowsPhone_8_1 = /Windows Phone 8.1/.test(navigator.userAgent); // Windows Phone 8.1
-//var isIE = isWindows || isWP8 || isWindowsPhone_8_1;
-var isIE = isWindows || isWP8;
-var isWebKit = !isIE; // TBD [Android or iOS]
+var isWindows = /Windows /.test(navigator.userAgent); // Windows
+var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
 // NOTE: In the core-master branch there is no difference between the default
 // implementation and implementation #2. But the test will also apply
@@ -25,12 +19,11 @@ var scenarioList = [
 
 var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
 
-// simple tests:
 var mytests = function() {
 
   for (var i=0; i<scenarioCount; ++i) {
 
-    describe(scenarioList[i] + ': Basic sql tx test(s)', function() {
+    describe(scenarioList[i] + ': tx sql results test(s)', function() {
       var scenarioName = scenarioList[i];
       var suiteName = scenarioName + ': ';
       var isWebSql = (i === 1);
@@ -54,9 +47,9 @@ var mytests = function() {
         }
       }
 
-      it(suiteName + 'US-ASCII String manipulation test',
+      it(suiteName + 'US-ASCII String manipulation results test',
         function(done) {
-          var db = openDatabase('ASCII-string-test.db', '1.0', 'Test', DEFAULT_SIZE);
+          var db = openDatabase('ASCII-string-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
           expect(db).toBeDefined();
 
@@ -73,12 +66,15 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-      // Only test ICU-UNICODE with Android 5.0(+) (Web SQL):
-      if (isWebSql && /Android [5-9]/.test(navigator.userAgent))
-        it(suiteName + 'ICU-UNICODE string manipulation test', function(done) {
-          if ((!isWebSql) && isAndroid) pending('BROKEN for Android version of plugin [with sqlite-connector]');
+      // NOTE: This test is ONLY for the following Android scenarios:
+      // 1. Web SQL test (Android 5.x/+)
+      // 2. Cordova-sqlcipher-adapter version of this plugin
+      if (isAndroid)
+        it(suiteName + 'Android ICU-UNICODE string manipulation test', function(done) {
+          if (isWebSql && /Android [1-4]/.test(navigator.userAgent)) pending('SKIP for Android versions 1.x-4.x Web SQL');
+          if (!isWebSql) pending('SKIP for plugin');
 
-          var db = openDatabase('UNICODE-string-test.db', '1.0', 'Test', DEFAULT_SIZE);
+          var db = openDatabase('ICU-UNICODE-string-manipulation-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
           expect(db).toBeDefined();
 
@@ -96,9 +92,9 @@ var mytests = function() {
           });
         });
 
-        it(suiteName + 'Simple INSERT test: check insertId & rowsAffected in result', function(done) {
+        it(suiteName + 'Simple INSERT results test: check insertId & rowsAffected in result', function(done) {
 
-          var db = openDatabase('INSERT-test.db', '1.0', 'Test', DEFAULT_SIZE);
+          var db = openDatabase('INSERT-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
           expect(db).toBeDefined();
 
@@ -119,9 +115,9 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-      it(suiteName + 'basic db sql transaction test',
-        function(done) {
-          var db = openDatabase('basic-db-sql-tx-test.db', '1.0', 'Test', DEFAULT_SIZE);
+        it(suiteName + 'basic db sql transaction results test', function(done) {
+
+          var db = openDatabase('basic-db-sql-tx-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
           expect(db).toBeDefined();
 
@@ -219,8 +215,7 @@ var mytests = function() {
 
         }, MYTIMEOUT);
 
-      it(suiteName + 'db transaction result object lifetime',
-        function(done) {
+        it(suiteName + 'db transaction result object lifetime', function(done) {
           var db = openDatabase('db-tx-result-lifetime-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
           expect(db).toBeDefined();
@@ -389,6 +384,212 @@ var mytests = function() {
           // not check_counted:
           //}, function() {
           //  console.log('first tx success cb OK');
+          });
+
+        }, MYTIMEOUT);
+
+        it(suiteName + 'tx sql starting with extra space results test', function(done) {
+          if (isWP8) pending('BROKEN for WP8');
+
+          var db = openDatabase('tx-sql-starting-with-extra-space-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
+
+          expect(db).toBeDefined();
+
+          var check_count = 0;
+
+          db.transaction(function(tx) {
+            // first tx object:
+            expect(tx).toBeDefined();
+
+            tx.executeSql('DROP TABLE IF EXISTS test_table');
+            tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data text, data_num integer)');
+
+            tx.executeSql(' INSERT INTO test_table (data, data_num) VALUES (?,?)', ['test', 100], function(tx, res) {
+              // check tx & res object parameters:
+              expect(tx).toBeDefined();
+              expect(res).toBeDefined();
+
+              expect(res.insertId).toBeDefined();
+              expect(res.rowsAffected).toBe(1);
+
+              db.transaction(function(tx) {
+                // second tx object:
+                expect(tx).toBeDefined();
+
+                tx.executeSql(' SELECT COUNT(id) AS cnt FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).cnt).toBe(1);
+                });
+
+                tx.executeSql('  SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(100);
+                });
+
+                tx.executeSql('   SELECT data FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data).toBe('test');
+                });
+
+                tx.executeSql('  UPDATE test_table SET data_num = ? WHERE data_num = 100', [101], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rowsAffected).toBe(1);
+                });
+
+                tx.executeSql('   SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(101);
+                });
+
+                tx.executeSql(" DELETE FROM test_table WHERE data LIKE 'tes%'", [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rowsAffected).toBe(1);
+                });
+
+                tx.executeSql('  SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(0);
+                });
+
+              }, function(e) {
+                // not expected:
+                expect(false).toBe(true);
+                expect(JSON.stringify(e).toBe('---'));
+                done();
+              }, function() {
+                console.log('second tx ok success cb');
+                expect(check_count).toBe(7);
+
+                done();
+              });
+
+            }, function(e) {
+              // not expected:
+              expect(false).toBe(true);
+              expect(JSON.stringify(e).toBe('---'));
+              done();
+            });
+          }, function(e) {
+            // not expected:
+            expect(false).toBe(true);
+            expect(JSON.stringify(e).toBe('---'));
+            done();
+          });
+
+        }, MYTIMEOUT);
+
+        it(suiteName + 'tx sql starting with extra semicolon results test', function(done) {
+          if (isWP8) pending('BROKEN for WP8');
+
+          // XXX [BUG #458] BROKEN for android.database
+          // NOTE: This is NOT broken when using Android-sqlite-connector
+          if (isAndroid && !isWebSql) pending('SKIP for Android version of plugin');
+
+          var db = openDatabase('tx-sql-starting-with-extra-semicolon-results-test.db', '1.0', 'Test', DEFAULT_SIZE);
+
+          expect(db).toBeDefined();
+
+          var check_count = 0;
+
+          db.transaction(function(tx) {
+            // first tx object:
+            expect(tx).toBeDefined();
+
+            tx.executeSql(';DROP TABLE IF EXISTS test_table');
+            tx.executeSql('; CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data text, data_num integer)');
+
+            tx.executeSql(';INSERT INTO test_table (data, data_num) VALUES (?,?)', ['test', 100], function(tx, res) {
+              // check tx & res object parameters:
+              expect(tx).toBeDefined();
+              expect(res).toBeDefined();
+
+              expect(res.insertId).toBeDefined();
+              expect(res.rowsAffected).toBe(1);
+
+              db.transaction(function(tx) {
+                // second tx object:
+                expect(tx).toBeDefined();
+
+                tx.executeSql(';SELECT COUNT(id) AS cnt FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).cnt).toBe(1);
+                });
+
+                tx.executeSql('; SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(100);
+                });
+
+                tx.executeSql(';; SELECT data FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data).toBe('test');
+                });
+
+                tx.executeSql('; ;UPDATE test_table SET data_num = ? WHERE data_num = 100', [101], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rowsAffected).toBe(1);
+                });
+
+                tx.executeSql(' ; SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(101);
+                });
+
+                tx.executeSql(";DELETE FROM test_table WHERE data LIKE 'tes%'", [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rowsAffected).toBe(1);
+                });
+
+                tx.executeSql('; SELECT data_num FROM test_table;', [], function(tx, res) {
+                  ++check_count;
+
+                  expect(res.rows.length).toBe(0);
+                });
+
+              }, function(e) {
+                // not expected:
+                expect(false).toBe(true);
+                expect(JSON.stringify(e).toBe('---'));
+                done();
+              }, function() {
+                console.log('second tx ok success cb');
+                expect(check_count).toBe(7);
+
+                done();
+              });
+
+            }, function(e) {
+              // not expected:
+              expect(false).toBe(true);
+              expect(JSON.stringify(e).toBe('---'));
+              done();
+            });
+          }, function(e) {
+            // not expected:
+            expect(false).toBe(true);
+            expect(JSON.stringify(e).toBe('---'));
+            done();
           });
 
         }, MYTIMEOUT);