+++ spec/www/spec/tx-semantics-test.js
@@ -30,7 +30,8 @@ function start(n) {
   if (wait == 0) test_it_done();
 }
 
-var isAndroid = /Android/.test(navigator.userAgent);
+var isWindows = /Windows /.test(navigator.userAgent); // Windows 8.1/Windows Phone 8.1/Windows 10
+var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
 // NOTE: In the core-master branch there is no difference between the default
 // implementation and implementation #2. But the test will also apply
@@ -477,8 +478,7 @@ var mytests = function() {
             db.transaction(function(tx) {
               tx.executeSql("insert into test_table (data, data_num) VALUES (?,?)", ['test', null], function(tx, res) {
                 expect(res).toBeDefined();
-                //if (!isWindows) // XXX TODO
-                  expect(res.rowsAffected).toEqual(1);
+                expect(res.rowsAffected).toEqual(1);
                 tx.executeSql("select * from bogustable", [], function(tx, res) {
                   ok(false, "select statement not supposed to succeed");
                 });
@@ -511,8 +511,7 @@ var mytests = function() {
               txg = tx;
               tx.executeSql("insert into test_table (data, data_num) VALUES (?,?)", ['test', null], function(tx, res) {
                 expect(res).toBeDefined();
-                //if (!isWindows) // XXX TODO
-                  expect(res.rowsAffected).toEqual(1);
+                expect(res.rowsAffected).toEqual(1);
               });
               start(1);
             }, function(err) {
@@ -534,51 +533,74 @@ var mytests = function() {
 
       });
 
-        test_it(suiteName + "readTransaction should throw on modification", function() {
-          stop();
-          var db = openDatabase("Database-readonly", "1.0", "Demo", DEFAULT_SIZE);
+        it(suiteName + "readTransaction should fail & report error on modification", function(done) {
+          var db = openDatabase("tx-readonly-test.db", "1.0", "Demo", DEFAULT_SIZE);
+
           db.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
-            tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (foo text)');
-            tx.executeSql('INSERT INTO test_table VALUES ("bar")');
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable1');
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable2');
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable3');
+
+            tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (data)');
+            tx.executeSql('INSERT INTO test_table VALUES (?)', ['first']);
+
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable1');
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable2');
+            tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable3');
           }, function () {}, function () {
             db.readTransaction(function (tx) {
               tx.executeSql('SELECT * from test_table', [], function (tx, res) {
                 equal(res.rows.length, 1);
-                equal(res.rows.item(0).foo, 'bar');
+                equal(res.rows.item(0).data, 'first');
               });
             }, function () {}, function () {
-              var tasks;
               var numDone = 0;
               var failed = false;
+              var tasks;
+
               function checkDone() {
                 if (++numDone === tasks.length) {
-                  start();
+                  done();
                 }
               }
               function fail() {
                 if (!failed) {
+                  expect(false).toBe(true);
+                  expect('readTransaction was supposed to fail').toBe('--');
                   failed = true;
-                  ok(false, 'readTransaction was supposed to fail');
 
-                  start();
+                  done();
                 }
               }
-              // all of these should throw an error
+
               tasks = [
+                // these transactions should be OK:
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql(' SELECT 1;');
+                  }, fail, checkDone);
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('; SELECT 1;');
+                  }, fail, checkDone);
+                },
+
+                // all of these transactions should report an error
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('UPDATE test_table SET foo = "another"');
+                  }, checkDone, fail);
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('INSERT INTO test_table VALUES ("another")');
+                  }, checkDone, fail);
+                },
                 function () {
                   db.readTransaction(function (tx) {
                     tx.executeSql('DELETE from test_table');
-                  }, checkDone, fail);
-                },
-                function () {
-                  db.readTransaction(function (tx) {
-                    tx.executeSql('UPDATE test_table SET foo = "baz"');
-                  }, checkDone, fail);
-                },
-                function () {
-                  db.readTransaction(function (tx) {
-                    tx.executeSql('INSERT INTO test_table VALUES ("baz")');
                   }, checkDone, fail);
                 },
                 function () {
@@ -588,9 +610,74 @@ var mytests = function() {
                 },
                 function () {
                   db.readTransaction(function (tx) {
-                    tx.executeSql('CREATE TABLE test_table2');
+                    // extra space before sql (OK)
+                    tx.executeSql(' CREATE TABLE test_table2 (data)');
                   }, checkDone, fail);
-                }
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    // two extra spaces before sql (OK)
+                    tx.executeSql('  CREATE TABLE test_table3 (data)');
+                  }, checkDone, fail);
+                },
+
+                // BUG #460:
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql(';  CREATE TABLE ExtraTestTable1 (data)');
+                  }, function(e) {
+                    // CORRECT
+                    if (!isWebSql) expect('Plugin FIXED, please update this test').toBe('--');
+                    checkDone();
+                  }, function() {
+                    // BUG #460: IGNORED for Plugin ONLY:
+                    if (!isWebSql) return checkDone(); // (returns undefined)
+                    expect(false).toBe(true);
+                    fail();
+                  });
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql(' ;  CREATE TABLE ExtraTestTable2 (data)');
+                  }, function(e) {
+                    // CORRECT
+                    if (!isWebSql) expect('Plugin FIXED, please update this test').toBe('--');
+                    checkDone();
+                  }, function() {
+                    // BUG #460: IGNORED for Plugin ONLY:
+                    if (!isWebSql) return checkDone(); // (returns undefined)
+                    expect(false).toBe(true);
+                    fail();
+                  });
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql(';CREATE TABLE ExtraTestTable3 (data)');
+                  }, function(e) {
+                    // CORRECT
+                    if (!isWebSql) expect('Plugin FIXED, please update this test').toBe('--');
+                    checkDone();
+                  }, function() {
+                    // BUG #460: IGNORED for Plugin ONLY:
+                    if (!isWebSql) return checkDone(); // (returns undefined)
+                    expect(false).toBe(true);
+                    fail();
+                  });
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql(';; CREATE TABLE ExtraTestTable4 (data)');
+                  }, function(e) {
+                    // CORRECT
+                    if (!isWebSql) expect('Plugin FIXED, please update this test').toBe('--');
+                    checkDone();
+                  }, function() {
+                    // BUG #460: IGNORED for Plugin ONLY:
+                    if (!isWebSql) return checkDone(); // (returns undefined)
+                    expect(false).toBe(true);
+                    fail();
+                  });
+                },
               ];
               for (var i = 0; i < tasks.length; i++) {
                 tasks[i]();